myeduca/src/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-braces==1.14.0
django-embed-video==1.3.3
djangorestframework==3.11.1
idna==2.10
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
